json: make methods static
