+++ meson.build
@@ -226,7 +226,9 @@ sources = [
   'src/global/vars.c',
   'src/global/vars_platform.c',
   'src/init.c',
-  'src/json.c',
+  'src/json/json_base.c',
+  'src/json/json_parse.c',
+  'src/json/json_write.c',
   'src/log.c',
   'src/memory.c',
   'src/specific/s_audio.c',
